Version 2.0.0 released
> JS code has been deeply rewritten. It is more flexible and should be
compatible with other JS codes that bind functions on the messageContent
element

> QuoteME supports now Touch Screen devices:
# Notebooks such as Sony Vaio
- With v1.x of this addon, any touch screen devices were rejected, so it
was not possible to use it on systems that have both a touch screen & a
mouse
- On these devices, the quoteME management hasn't changed (the text
range selection is not possible anyway)

# Mobile devices such as Ipad, Ipod
- To use this addon with mobile devices, it requires a recent operative
system & a recent browser that support the css fixed position (it seems
there were problems before)
- I'm not going to support this function. The documentation is not easy
to find. I've just tested this on iOS & on Win7 with Chrome and its
touch events emulation. So if it works, good for you, otherwise sorry.
- On these devices, the quoteME popup has been changed into a horizontal
button that will appear at the bottom of the visible screen (which
explains why the fixed position must be supported)

Please note that on any touch screen devices, the "After click" mode
will be automatically activated. This is done on purpose. If one of
these devices doesn't support QuoteME, then your users will not have any
problems. If you want to enable the same solution with mobile devices
that don't have a touch screen, you will have to install the "Browser
Detection" addon : http://xenforo.com/community/resources/1098/.

Important: on iOS, when some quotes are inserted inside the editor, the
editor is focused and the OS keyboard is displayed. Don't try to
continue to select some quotes while this keyboard is opened, it will
not work and will automatically refocused the editor. You must first
close the keyboard, then it will work. For this reason, the TransMode is
recommended. It will allow you to insert all quotes at once.

> New option to disable the QuoteME addon on mobile devices (which can
be different from touch screen devices)

> Fix the QuoteME TinyMCE Quattro Plugin: a click on the QuoteME top
activation trigger will also update the MCE4 button state